+++ src/atlas/actions/BuildStatistics.cc
@@ -265,7 +265,7 @@ void build_statistics( Mesh& mesh )
     ArrayView<double,1> dual_volumes ( nodes.field("dual_volumes") );
     ArrayView<double,1> dual_delta_sph  ( nodes.create_field<double>( "dual_delta_sph", 1 ) );
 
-    for( int jnode=0; jnode<nodes.shape(0); ++jnode )
+    for(size_t jnode = 0; jnode < nodes.shape(0); ++jnode)
     {
       const double lat = lonlat(jnode,LAT)*DEG_TO_RAD;
       const double hx = radius_km*std::cos(lat)*DEG_TO_RAD;
@@ -275,7 +275,7 @@ void build_statistics( Mesh& mesh )
 
     if( eckit::mpi::size() == 1 )
     {
-      for( int jnode=0; jnode<nodes.shape(0); ++jnode )
+      for(size_t jnode = 0; jnode < nodes.shape(0); ++jnode)
       {
         ofs << std::setw(idt) << dual_delta_sph(jnode)
             << "\n";
